+++ lib/src/controls/navigation/navigation_view/indicators.dart
@@ -45,7 +45,7 @@ class NavigationIndicator extends StatefulWidget {
 }
 
 class NavigationIndicatorState<T extends NavigationIndicator> extends State<T> {
-  List<Offset>? offsets;
+  Iterable<Offset>? offsets;
 
   @override
   void initState() {
@@ -70,7 +70,7 @@ class NavigationIndicatorState<T extends NavigationIndicator> extends State<T> {
   }
 
   NavigationPane get pane {
-    return InheritedNavigationView.of(context).pane!;
+    return _InheritedNavigationView.of(context).pane!;
   }
 
   int get selectedIndex {
@@ -82,7 +82,7 @@ class NavigationIndicatorState<T extends NavigationIndicator> extends State<T> {
   }
 
   Axis get axis {
-    if (InheritedNavigationView.maybeOf(context)?.displayMode ==
+    if (_InheritedNavigationView.maybeOf(context)?.displayMode ==
         PaneDisplayMode.top) {
       return Axis.vertical;
     }
@@ -90,11 +90,11 @@ class NavigationIndicatorState<T extends NavigationIndicator> extends State<T> {
   }
 
   int get itemIndex {
-    return InheritedNavigationView.of(context).currentItemIndex;
+    return _InheritedNavigationView.of(context).currentItemIndex;
   }
 
-  int get oldIndex {
-    return InheritedNavigationView.of(context).oldIndex;
+  int get previousItemIndex {
+    return _InheritedNavigationView.of(context).previousItemIndex;
   }
 
   PaneItem get item {
@@ -180,7 +180,8 @@ class StickyNavigationIndicator extends NavigationIndicator {
     super.color,
     super.duration,
     this.topPadding = 12.0,
-    this.leftPadding = 10.0,
+    this.leftPadding = kPaneItemMinHeight * 0.3,
+    this.indicatorSize = 2.75,
   });
 
   /// The padding used on both horizontal sides of the indicator when the
@@ -194,6 +195,14 @@ class StickyNavigationIndicator extends NavigationIndicator {
   ///
   /// Defaults to 10.0
   final double leftPadding;
+
+  /// The size of the indicator.
+  ///
+  /// On top display mode, this represents the height of the indicator. On other
+  /// display modes, this represents the width of the indicator.
+  ///
+  /// Defaults to 2.0
+  final double indicatorSize;
 
   @override
   NavigationIndicatorState<StickyNavigationIndicator> createState() =>
@@ -245,11 +254,11 @@ class _StickyNavigationIndicatorState
     if (itemIndex.isNegative) return false;
 
     if (itemIndex == selectedIndex) return true;
-    return itemIndex == oldIndex && _old != oldIndex;
+    return itemIndex == previousItemIndex && _old != previousItemIndex;
   }
 
-  bool get isAbove => oldIndex < selectedIndex;
-  bool get isBelow => oldIndex > selectedIndex;
+  bool get isAbove => previousItemIndex < selectedIndex;
+  bool get isBelow => previousItemIndex > selectedIndex;
 
   @override
   void didChangeDependencies() {
@@ -264,13 +273,13 @@ class _StickyNavigationIndicatorState
 
     _old = (PageStorage.of(context).readState(
           context,
-          identifier: 'oldIndex$itemIndex',
+          identifier: 'previousItemIndex$itemIndex',
         ) as num?)
             ?.toInt() ??
         _old;
 
     // do not perform the animation twice
-    if (_old == oldIndex) {
+    if (_old == previousItemIndex) {
       return;
     }
 
@@ -312,12 +321,12 @@ class _StickyNavigationIndicatorState
       }
     }
 
-    _old = oldIndex;
+    _old = previousItemIndex;
     if (mounted) {
       PageStorage.of(context).writeState(
         context,
         _old,
-        identifier: 'oldIndex$itemIndex',
+        identifier: 'previousItemIndex$itemIndex',
       );
       setState(() {});
     }
@@ -350,11 +359,17 @@ class _StickyNavigationIndicatorState
           child: isHorizontal
               ? Align(
                   alignment: AlignmentDirectional.centerStart,
-                  child: Container(width: 2.5, decoration: decoration),
+                  child: Container(
+                    width: widget.indicatorSize,
+                    decoration: decoration,
+                  ),
                 )
               : Align(
                   alignment: AlignmentDirectional.bottomCenter,
-                  child: Container(height: 2.5, decoration: decoration),
+                  child: Container(
+                    height: widget.indicatorSize,
+                    decoration: decoration,
+                  ),
                 ),
           builder: (context, child) {
             if (!isSelected) {
@@ -367,7 +382,7 @@ class _StickyNavigationIndicatorState
               padding: isHorizontal
                   ? EdgeInsetsDirectional.only(
                       start: () {
-                        final x = offsets![itemIndex].dx;
+                        final x = offsets!.elementAt(itemIndex).dx;
                         if (parent != null) {
                           final isOpen =
                               parent!.expanderKey.currentState?._open ?? false;
@@ -377,7 +392,7 @@ class _StickyNavigationIndicatorState
 
                           final parentIndex =
                               pane.effectiveItems.indexOf(parent!);
-                          final parentX = offsets![parentIndex].dx;
+                          final parentX = offsets!.elementAt(parentIndex).dx;
                           return parentX;
                         }
                         return x;