+++ lib/src/controls/navigation/navigation_view/pane.dart
@@ -665,7 +665,7 @@ class _TopNavigationPaneState extends State<_TopNavigationPane> {
   @override
   Widget build(BuildContext context) {
     assert(debugCheckHasFluentTheme(context));
-    final view = InheritedNavigationView.of(context);
+    final view = _InheritedNavigationView.of(context);
     final height = widget.pane.size?.topHeight ?? kOneLineTileHeight;
     return SizedBox(
       key: widget.pane.paneKey,
@@ -698,11 +698,11 @@ class _TopNavigationPaneState extends State<_TopNavigationPane> {
                     placementMode: FlyoutPlacementMode.bottomCenter,
                     forceAvailableSpace: true,
                     builder: (context) {
-                      return InheritedNavigationView(
+                      return _InheritedNavigationView(
                         displayMode: view.displayMode,
                         currentItemIndex: view.currentItemIndex,
                         minimalPaneOpen: view.minimalPaneOpen,
-                        oldIndex: view.oldIndex,
+                        previousItemIndex: view.previousItemIndex,
                         pane: view.pane,
                         child: MenuFlyout(
                           items: _localItemHold
@@ -826,7 +826,7 @@ class _MenuFlyoutPaneItem extends MenuFlyoutItemBase {
     final size = Flyout.of(context).size;
     final theme = NavigationPaneTheme.of(context);
     final fluentTheme = FluentTheme.of(context);
-    final view = InheritedNavigationView.of(context);
+    final view = _InheritedNavigationView.of(context);
 
     final selected = view.pane?.isSelected(item) ?? false;
     final titleText = item.title?.getProperty<String>() ?? '';
@@ -1037,6 +1037,7 @@ class _CompactNavigationPane extends StatelessWidget {
     this.listKey,
     this.onToggle,
     this.onOpenSearch,
+    this.onAnimationEnd,
   }) : super(key: pane.key);
 
   final NavigationPane pane;
@@ -1044,9 +1045,11 @@ class _CompactNavigationPane extends StatelessWidget {
   final GlobalKey? listKey;
   final VoidCallback? onToggle;
   final VoidCallback? onOpenSearch;
+  final VoidCallback? onAnimationEnd;
 
-  Widget _buildItem(BuildContext context, NavigationPaneItem item) {
+  static Widget _buildItem(BuildContext context, NavigationPaneItem item) {
     assert(debugCheckHasFluentTheme(context));
+    final pane = _InheritedNavigationView.of(context).pane!;
     if (item is PaneItemHeader) {
       // Item Header is not visible on compact pane
       return const SizedBox();
@@ -1091,6 +1094,7 @@ class _CompactNavigationPane extends StatelessWidget {
       duration: theme.animationDuration ?? Duration.zero,
       curve: theme.animationCurve ?? Curves.linear,
       width: pane.size?.compactWidth ?? kCompactNavigationPaneWidth,
+      onEnd: onAnimationEnd,
       child: Align(
         key: pane.paneKey,
         alignment: AlignmentDirectional.topCenter,
@@ -1104,7 +1108,6 @@ class _CompactNavigationPane extends StatelessWidget {
                 pane,
                 onPressed: () {
                   onToggle?.call();
-                  onOpenSearch?.call();
                 },
                 padding: showReplacement ? EdgeInsets.zero : topPadding,
               );
@@ -1123,6 +1126,7 @@ class _CompactNavigationPane extends StatelessWidget {
                 false,
                 () {
                   onToggle?.call();
+                  onOpenSearch?.call();
                 },
               ),
             ),
@@ -1159,6 +1163,7 @@ class _OpenNavigationPane extends StatefulWidget {
     this.onToggle,
     this.onItemSelected,
     this.initiallyOpen = false,
+    this.onAnimationEnd,
   }) : super(key: pane.key);
 
   final NavigationPane pane;
@@ -1168,13 +1173,18 @@ class _OpenNavigationPane extends StatefulWidget {
   final VoidCallback? onItemSelected;
   final NavigationPaneThemeData theme;
   final bool initiallyOpen;
+  final VoidCallback? onAnimationEnd;
 
   static Widget buildItem(
     BuildContext context,
     NavigationPane pane,
     NavigationPaneItem item, [
     VoidCallback? onChanged,
+    double? width,
   ]) {
+    if (width != null && width < kOpenNavigationPaneWidth / 1.5) {
+      return _CompactNavigationPane._buildItem(context, item);
+    }
     if (item is PaneItemHeader) {
       return item.build(context);
     } else if (item is PaneItemSeparator) {
@@ -1227,16 +1237,12 @@ class _OpenNavigationPaneState extends State<_OpenNavigationPane>
       vsync: this,
       duration: theme.animationDuration,
     );
-    if (widget.initiallyOpen) {
-      controller.value = 1;
-    } else {
-      controller.forward();
-      PageStorage.of(context).writeState(
-        context,
-        true,
-        identifier: 'openModeOpen',
-      );
-    }
+    controller.forward(from: widget.initiallyOpen ? 1 : 0);
+    PageStorage.of(context).writeState(
+      context,
+      true,
+      identifier: 'openModeOpen',
+    );
   }
 
   @override
@@ -1271,17 +1277,15 @@ class _OpenNavigationPaneState extends State<_OpenNavigationPane>
       paneHeaderHeight = -1.0;
     }
 
-    return SizeTransition(
-      axisAlignment: -1,
-      axis: Axis.horizontal,
-      sizeFactor: Tween<double>(begin: 0, end: 1.0).animate(CurvedAnimation(
-        parent: controller,
-        curve: theme.animationCurve ?? Curves.linear,
-      )),
-      child: SizedBox(
-        key: widget.paneKey,
-        width: paneWidth,
-        child: Column(
+    return AnimatedContainer(
+      duration: theme.animationDuration ?? Duration.zero,
+      curve: theme.animationCurve ?? Curves.linear,
+      key: widget.paneKey,
+      width: paneWidth,
+      onEnd: widget.onAnimationEnd,
+      child: LayoutBuilder(builder: (context, constraints) {
+        final width = constraints.maxWidth;
+        return Column(
           crossAxisAlignment: CrossAxisAlignment.start,
           key: widget.pane.paneKey,
           children: [
@@ -1304,6 +1308,7 @@ class _OpenNavigationPaneState extends State<_OpenNavigationPane>
                             ),
                             child: DefaultTextStyle.merge(
                               style: theme.itemHeaderTextStyle,
+                              maxLines: 1,
                               child: widget.pane.header!,
                             ),
                           ),
@@ -1316,13 +1321,28 @@ class _OpenNavigationPaneState extends State<_OpenNavigationPane>
                 }(),
               ),
             if (widget.pane.autoSuggestBox != null)
-              Container(
-                padding: theme.iconPadding ?? EdgeInsets.zero,
-                height: 41.0,
-                alignment: AlignmentDirectional.center,
-                margin: topPadding,
-                child: widget.pane.autoSuggestBox!,
-              ),
+              if (width > kOpenNavigationPaneWidth / 1.5)
+                Container(
+                  padding: theme.iconPadding ?? EdgeInsets.zero,
+                  height: 41.0,
+                  alignment: AlignmentDirectional.center,
+                  margin: topPadding,
+                  child: widget.pane.autoSuggestBox!,
+                )
+              else
+                Padding(
+                  padding: topPadding,
+                  child: PaneItem(
+                    title: Text(FluentLocalizations.of(context).clickToSearch),
+                    icon: widget.pane.autoSuggestBoxReplacement!,
+                    body: const SizedBox.shrink(),
+                  ).build(
+                    context,
+                    false,
+                    () {},
+                    displayMode: PaneDisplayMode.compact,
+                  ),
+                ),
             Expanded(
               child: ListView(
                 shrinkWrap: true,
@@ -1334,6 +1354,7 @@ class _OpenNavigationPaneState extends State<_OpenNavigationPane>
                     widget.pane,
                     item,
                     widget.onItemSelected,
+                    width,
                   );
                 }).toList(),
               ),
@@ -1348,12 +1369,13 @@ class _OpenNavigationPaneState extends State<_OpenNavigationPane>
                   widget.pane,
                   item,
                   widget.onItemSelected,
+                  width,
                 );
               }).toList(),
             ),
           ],
-        ),
-      ),
+        );
+      }),
     );
   }
 }