+++ lib/src/controls/navigation/navigation_view/style.dart
@@ -83,16 +83,21 @@ class NavigationPaneTheme extends InheritedTheme {
 /// The theme data used by [NavigationView]. The default theme
 /// data used is [NavigationPaneThemeData.standard].
 class NavigationPaneThemeData with Diagnosticable {
-  /// The pane background color. If null, [FluentThemeData.micaBackgroundColor]
-  /// is used.
+  /// The pane background color.
   final Color? backgroundColor;
 
-  /// The color of the tiles. If null, [ButtonThemeData.uncheckedInputColor]
-  /// is used
+  /// The pane background color when there is an overlay, such as minimal or
+  /// compact display modes.
+  final Color? overlayBackgroundColor;
+
+  /// The color of the tiles.
+  ///
+  /// If null, [ButtonThemeData.uncheckedInputColor] is used
   final ButtonState<Color?>? tileColor;
 
-  /// The highlight color used on the tiles. If null, [FluentThemeData.accentColor]
-  /// is used.
+  /// The highlight color used on the tiles.
+  ///
+  /// If null, [FluentThemeData.accentColor] is used.
   final Color? highlightColor;
 
   final EdgeInsetsGeometry? labelPadding;
@@ -115,6 +120,7 @@ class NavigationPaneThemeData with Diagnosticable {
 
   const NavigationPaneThemeData({
     this.backgroundColor,
+    this.overlayBackgroundColor,
     this.tileColor,
     this.highlightColor,
     this.labelPadding,
@@ -144,6 +150,7 @@ class NavigationPaneThemeData with Diagnosticable {
       animationDuration: animationDuration,
       animationCurve: animationCurve,
       backgroundColor: resources.solidBackgroundFillColorBase,
+      overlayBackgroundColor: resources.systemFillColorSolidNeutralBackground,
       highlightColor: highlightColor,
       itemHeaderTextStyle: typography.bodyStrong,
       selectedTextStyle: ButtonState.resolveWith((states) {
@@ -201,6 +208,8 @@ class NavigationPaneThemeData with Diagnosticable {
           EdgeInsetsGeometry.lerp(a?.headerPadding, b?.headerPadding, t),
       tileColor: ButtonState.lerp(a?.tileColor, b?.tileColor, t, Color.lerp),
       backgroundColor: Color.lerp(a?.backgroundColor, b?.backgroundColor, t),
+      overlayBackgroundColor:
+          Color.lerp(a?.overlayBackgroundColor, b?.overlayBackgroundColor, t),
       itemHeaderTextStyle:
           TextStyle.lerp(a?.itemHeaderTextStyle, b?.itemHeaderTextStyle, t),
       selectedTextStyle: ButtonState.lerp(
@@ -229,6 +238,8 @@ class NavigationPaneThemeData with Diagnosticable {
       headerPadding: style?.headerPadding ?? headerPadding,
       tileColor: style?.tileColor ?? tileColor,
       backgroundColor: style?.backgroundColor ?? backgroundColor,
+      overlayBackgroundColor:
+          style?.overlayBackgroundColor ?? overlayBackgroundColor,
       itemHeaderTextStyle: style?.itemHeaderTextStyle ?? itemHeaderTextStyle,
       selectedTextStyle: style?.selectedTextStyle ?? selectedTextStyle,
       unselectedTextStyle: style?.unselectedTextStyle ?? unselectedTextStyle,
@@ -249,6 +260,7 @@ class NavigationPaneThemeData with Diagnosticable {
     properties
       ..add(DiagnosticsProperty('tileColor', tileColor))
       ..add(ColorProperty('backgroundColor', backgroundColor))
+      ..add(ColorProperty('overlayBackgroundColor', overlayBackgroundColor))
       ..add(ColorProperty('highlightColor', highlightColor))
       ..add(
           DiagnosticsProperty<EdgeInsetsGeometry>('labelPadding', labelPadding))